Made get_logical_form handle reverse
